requirements.txt
langchain
langchain_community
langchain_core
langchain_google_genai
langchain_tavily
google-generativeai
faiss-cpu
fastapi
pypdf
uvicorn
python-multipart
langgraph
pymongo
passlib[bcrypt]
python-jose
